refactor(ref-api-02): slim subscription_controller 400 → 219 LOC (PR 1/7)
Extract webhook processing and serialisation from the controller so it
stays close to the ≤200 LOC target (REF-API-02 pattern):

- app/schemas/openapi/subscription/response.py (new): serialize_subscription()
  extracted verbatim; creates the app/schemas/openapi/ hierarchy.
- app/controllers/subscription_webhook_handler.py (new): _process_webhook_snapshot
  + handle_webhook_request() (full webhook POST body). All webhook-only imports
  move here; controller's handle_webhook() becomes a 1-line delegator.
- subscription_controller.py: 400 → 219 LOC. Re-exports
  _process_webhook_snapshot, _extract_event_id, _extract_provider_snapshot
  preserved for billing_webhooks_cli compatibility.

All 49 billing subscription tests pass. Contract unchanged.

Part of #1086.
